Can upload files to S3
